requirements_fixed.txt
# Core Python libraries for AI Trading System V2

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Technical analysis
TA-Lib>=0.4.25

# Machine Learning and AI
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.2.0
optuna>=3.1.0

# Deep Learning
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.20.0

# Trading platform integration
MetaTrader5>=5.0.45

# Database and caching
redis>=4.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration
PyYAML>=6.0
toml>=0.10.2
python-dotenv>=1.0.0

# Web framework and dashboard
streamlit>=1.25.0
plotly>=5.15.0
dash>=2.10.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Async programming
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0
asyncpg>=0.28.0

# Data validation
pydantic>=2.0.0
typing-extensions>=4.5.0

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Time series analysis
statsmodels>=0.14.0
arch>=6.0.0
pmdarima>=2.0.0

# Parallel processing
joblib>=1.3.0